修复集成测试发现的所有问题

+++ src/main/java/com/zhss/eshop/membership/service/impl/PayOrderMembershipUpdater.java
@@ -137,7 +137,7 @@ public class PayOrderMembershipUpdater extends AbstractMembershipUpdater<Object,
 	@Override
 	protected Map<String, Object> undoUpdatedMemberLevel(Long userAccountId, 
 			Object parameter) throws Exception {
-		Long totalOrderAmount = (Long) parameter;
+		Double totalOrderAmount = (Double) parameter;
 		
 		Map<String, Object> result = new HashMap<String, Object>(CollectionSize.DEFAULT);
 		
@@ -165,7 +165,7 @@ public class PayOrderMembershipUpdater extends AbstractMembershipUpdater<Object,
 	@Override
 	protected Map<String, Object> undoUpdatedMemberPoint(Long userAccountId, 
 			Object parameter) throws Exception {
-		Long totalOrderAmount = (Long) parameter;
+		Double totalOrderAmount = (Double) parameter;
 		
 		Map<String, Object> result = new HashMap<String, Object>(CollectionSize.DEFAULT);
